Refactor các ValueObject và repository, cập nhật entity và xử lý lỗi
- Chuyển Template* từ record sang class kế thừa ValueObject, thêm kiểm tra equality.
- Cập nhật các repository adapter dùng MongoHelper, thêm phương thức updateEntity.
- Xóa InfrastructureToHttpExceptionMapper và các enum liên quan, đơn giản hóa InfrastructureException.
- Thêm updateEntity trong các mapper để cập nhật entity mà không tạo mới.
- Refactor RemoteApiCaller ném HttpException thay vì RemoteException.

+++ notification-service/domain/src/main/java/studydocs/notification/domain/aggregate/FcmToken.java
@@ -1,6 +1,6 @@
 package studydocs.notification.domain.aggregate;
 
-import io.github.domain.aggregate.base.AggregateRoot;
+import io.github.domain.aggregate.AggregateRoot;
 import studydocs.notification.domain.vo.FcmTokenValue;
 
 import java.util.UUID;
@@ -10,7 +10,7 @@ public class FcmToken extends AggregateRoot {
     private UUID userId;
 
     /// Constructor
-    private FcmToken(UUID id, Integer version) {
+    private FcmToken(UUID id, long version) {
         super(id,version);
     }
 
@@ -19,7 +19,7 @@ public class FcmToken extends AggregateRoot {
     }
 
     /// Factory method
-    public static FcmToken reconstruct(UUID id, Integer version, UUID userId, String value) {
+    public static FcmToken reconstruct(UUID id, long version, UUID userId, String value) {
         var entity = new FcmToken(id,version);
         entity.userId = userId;
         entity.value = new FcmTokenValue(value);
